squashed some bugs and added some basic documentation
Effects, particularly those with Holds,
were not being processed correctly, and that seems to be resolved,
now.

added tested examples, which helped show some of the bugs.

added some methods (like blink()) to LedWriter, but should probably make some of the
stuff protected or private for cleanliness; I like the idea
of everything staying public, but it's a bit messy.